Added attribute handlers to Task

+++ database/migrations/2021_03_12_212827_create_task_statuses_dictionary.php
@@ -14,7 +14,7 @@ class CreateTaskStatusesDictionary extends Migration
     public function up()
     {
         Schema::create('dic_task_statuses', function (Blueprint $table) {
-            $table->unsignedInteger('id')->unique();
+            $table->unsignedBigInteger('id')->unique();
             $table->string('status');
         });
     }
